Fixed NaN error

+++ lib/GymClass/InstructorClassDetails.dart
@@ -703,8 +703,13 @@ class InstructorClassDetailsState extends State<InstructorClassDetails> {
 
     if (response.statusCode == 200) {
       GetResponse res = GetResponse.fromJson(json.decode(response.body));
-      int temp = ((res.ratingSum / res.ratingCount) * 10).truncate();
-      instructorRating = (temp) / 10;
+      if(res.ratingCount == 0){
+        instructorRating = 0;
+      }
+      else{
+        int temp = ((res.ratingSum / res.ratingCount) * 10).truncate();
+        instructorRating = (temp) / 10;
+      }
       return response.body;
     } else {
       throw Exception('Failed to retrieve user data. Please try again later');
@@ -723,11 +728,17 @@ class InstructorClassDetailsState extends State<InstructorClassDetails> {
     String url =
         'https://gymmoveswebapi.azurewebsites.net/api/ratings/class?classid=$classID';
     Response response = await get(url);
+    print(response.body);
 
     if (response.statusCode == 200) {
       GetResponse res = GetResponse.fromJson(json.decode(response.body));
-      int temp = ((res.ratingSum / res.ratingCount) * 10).truncate();
-      classRating = (temp) / 10;
+      if(res.ratingCount == 0){
+        classRating = 0;
+      }
+      else{
+        int temp = ((res.ratingSum / res.ratingCount) * 10).truncate();
+        classRating = (temp) / 10;
+      }
       return response.body;
     } else {
       throw Exception('Failed to retrieve user data. Please try again later');
